MemberId Int형으로 저장되도록 통일

+++ app/src/main/java/com/somasoma/wagglewaggle/data/model/dto/auth/SignUpResponse.kt
@@ -2,7 +2,7 @@ package com.somasoma.wagglewaggle.data.model.dto.auth
 
 data class SignUpResponse(
     val isNewMember: String?,
-    val memberId: Long?,
+    val memberId: Int?,
     val accessToken: String?,
     val refreshToken: String?,
     val grantType: String?,
