+++ app/src/main/java/com/somasoma/wagglewaggle/domain/usecase/member/GetMemberUseCase.kt
@@ -4,5 +4,5 @@ import com.somasoma.wagglewaggle.domain.repository.MemberRepository
 import javax.inject.Inject
 
 class GetMemberUseCase @Inject constructor(private val memberRepository: MemberRepository) {
-    suspend fun getMember(memberId: Long) = memberRepository.getMember(memberId)
+    suspend fun getMember(memberId: Int) = memberRepository.getMember(memberId)
 }
